+++ data/settings.js
@@ -177,6 +177,6 @@ const SETTINGS = {
 };
  const APP_META = {
   name: "CTA-Manager Lite",
-  version: "v0.13.5"
+  version: "v0.13.6"
 };
 window.APP_META = APP_META;